# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "海外算力服务",
    "description": "海外算力服务",
    "version": "0.0.1"
  },
  "host": "192.168.1.36:8999",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/hashrate/admin/auth/v1/login/password": {
      "post": {
        "summary": "账号密码登录",
        "operationId": "VerifyLoginPassword",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/VerifyLoginCodeResp"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyLoginPasswordReq"
            }
          }
        ],
        "tags": [
          "admin/auth"
        ]
      }
    },
...
  },
  "definitions": {
...
    "VerifyLoginCodeResp": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "description": " JWT令牌"
        }
      },
      "title": "VerifyLoginCodeResp",
      "required": [
        "token"
      ]
    },
    "VerifyLoginPasswordReq": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "description": " 登录邮箱"
        },
        "password": {
          "type": "string",
          "description": " 登录密码"
        },
        "mfa_code": {
          "type": "string",
          "description": " MFA验证码"
        }
      },
      "title": "VerifyLoginPasswordReq",
      "required": [
        "email",
        "password"
      ]
    },
...
  },
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "description": "Enter JWT Bearer token **_only_**",
      "name": "Authorization",
      "in": "header"
    }
  }
}